+++ data/cube/aggregate/Aggregator.js
@@ -28,4 +28,23 @@ export class Aggregator {
     replace(rows, currVal, update) {
         return this.aggregate(rows, update.field.name);
     }
+
+
+    /**
+     * Call function on all *leaf* children of a set of children.
+     *
+     * @param {Array} rows - array of child rows
+     * @param {function} fn - the function to call on each leaf.
+     *
+     * @returns {Array}
+     */
+    forEachLeaf(rows, fn) {
+        for (const row of rows) {
+            if (row.isLeaf) {
+                fn(row);
+            } else {
+                this.forEachLeaf(row.children, fn);
+            }
+        }
+    }
 }